requirements.txt
Abg==2.3.6
aiofiles
aiohttp>=2.2.5
httpx[http2]>=0.18.1,<1.0.0
alphabet_detector==0.0.7
beautifulsoup4==4.11.1
bleach==6.0.0
cachetools
configparser==5.3.0
dateparser
emoji==1.7.0
feedparser==6.0.10
ffmpeg-python==0.2.0
future==0.18.3
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython==3.1.31
markdown2==2.4.10
multicolorcaptcha==1.2.0
pillow==9.4.0
psutil==5.9.4
psycopg2-binary==2.9.7
pyYAML==6.0.1
pyrate-limiter==2.8.5
python-telegram-bot==13.15
requests==2.31.0
spamwatch==0.3.0
speedtest-cli==2.1.3
SQLAlchemy==1.4.44
telethon==1.30.3
ujson==5.8.0
pycryptodome==3.18.0
prettyconf
base58==2.1.1
pytimeparse==1.1.8
SibylSystem==0.0.24
TgCrypto
pyrogram==2.0.106
python-arq==6.0.5
python-dotenv==0.21.1
ffmpeg-python==0.2.0
redis==4.1.3
telegraph==2.2.0
motor
pymongo[srv]==4.5.0
odmantic==0.9.2
opencv-python-headless==4.8.0.76
pykeyboard==0.1.5
search_engine_parser==0.6.8
wget==3.2
wikipedia==1.4.0
bing_image_downloader==1.1.2
lyricsgenius==3.0.1
jikanpy==4.3.2
gtts==2.3.1
googletrans==2.4.0
gpytranslate==1.5.1
cloudscraper==1.2.70
carbonnow
pretty_errors
